moved shared folder from Dashboard to components
